requirements.txt
pandas
openpyxl
langchain
langchain-community
langchain-google-genai
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
tenacity
google-generativeai